moved error message to git package
